requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
pandas==2.2.0
numpy==1.26.4
plotly==5.18.0
streamlit==1.32.0
PyGithub==2.2.0
psutil==5.9.8
